Create activity5.c++
also, a light exercise. However, at the same time it is simple, it is complicated (for me, at least) fun fact, according to Sherlock Holmes the most complicated cases were the simplest ones, while the ones that seemed to be the most complexes ended up being the simplest ones.  If you ask me like, ok Arthur but how do you know that? ( U won't tho) I'll say that the reason for that is that I have a mammoth book in portuguese with thousands and a few more thousands of cases and storys of Sherlock Holmes, cool right? Ikik.